Separated tests into their own files
